requirements.txt
ccxt == 4.4.60
python-dotenv == 1.0.1
aiohttp == 3.9.5
python-binance==1.0.27